Delegation role controls added

+++ src/main/java/com/example/appent/service/DelegationService.java
@@ -4,8 +4,13 @@ import com.example.appent.entity.DelegationEntity;
 import com.example.appent.exception.DelegationEmpty;
 import com.example.appent.exception.DelegationExist;
 import com.example.appent.exception.DelegationNotExisting;
+import com.example.appent.exception.OrganisateurNotExisting;
+import com.example.appent.repository.ControlleurRepository;
 import com.example.appent.repository.DelegationRepository;
+import com.example.appent.repository.OrganisateurRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -14,27 +19,36 @@ import java.util.List;
 @Service
 public class DelegationService {
     private DelegationRepository delegationRepository;
+    private OrganisateurRepository organisateurRepository;
 
     @Autowired
-    public DelegationService(DelegationRepository delegationRepository) {
+    public DelegationService(DelegationRepository delegationRepository, OrganisateurRepository organisateurRepository) {
         this.delegationRepository = delegationRepository;
+        this.organisateurRepository = organisateurRepository;
     }
 
-    public DelegationEntity createDelegation(String nom) throws DelegationExist {
+    public DelegationEntity createDelegation(String nom, String mail) throws DelegationExist, OrganisateurNotExisting {
         DelegationEntity delegation = this.delegationRepository.findByNom(nom);
         if(delegation != null) {
             throw new DelegationExist();
+        }
+        if(this.organisateurRepository.findByEmail(mail) == null) {
+            throw new OrganisateurNotExisting("Organisateur n'existe pas");
         }
         delegation= new DelegationEntity(nom);
         return this.delegationRepository.save(delegation);
     }
 
-    public void DeleteDelegation(String nom) throws DelegationNotExisting {
+    public ResponseEntity<String> DeleteDelegation(String nom, String mail) throws DelegationNotExisting, OrganisateurNotExisting {
         DelegationEntity delegation = this.delegationRepository.findByNom(nom);
         if (delegation == null) {
             throw new DelegationNotExisting();
         }
+        if(this.organisateurRepository.findByEmail(mail) == null) {
+            throw new OrganisateurNotExisting("Organisateur n'existe pas");
+        }
         this.delegationRepository.delete(delegation);
+        return new ResponseEntity<>("Delegation " +nom +" supprimée", HttpStatus.OK);
     }
